backend/requirements.txt
Django==3.2.4
djangorestframework==3.12.4
fcm-django==0.3.10
django-cors-headers==3.7.0
